refactor(tests/docs): consolidate accumulator tests; scrub stale legacy references
- Merge the accumulator tests from the legacy-named test_metric_results_processor.py
  (which tested MetricsAccumulator, not the deleted MetricResultsProcessor) into
  test_metrics_accumulator.py and delete the misnamed file.
- Fix docs that still described the retired storage/aggregation model:
  list-metric-aggregation.md (MetricArray -> ColumnStore numpy column) and the
  metrics-flow diagram (per-instance aggregate_value fold -> aggregation_kind
  column fold).
